Added the choices_by_list to the context
